Event handler optimizes scroll function and added scroll to element
All event listeners get handled by the Event handler, which
prevents memory leaking caused by an event having chains of anonymous
functions if using so.

Scroll to element is now functional, by using the Listen function
on a parent element; it uses the bubbling phase
to target children that fired an event.